+++ amd/src/recommendationdetailsmodal.js
@@ -14,11 +14,10 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Theme Boost Union - JS code for course listing details modal.
+ * Theme Boost Union - JS code for recommendation details modal.
  *
- * @module     theme_boost_union/courselistingdetailsmodal
- * @copyright  2025 Alexander Bias, ssystems GmbH <abias@ssystems.de>
- *             based on core_admin/themeselector/preview_modal by David Woloszyn <david.woloszyn@moodle.com>
+ * @module     theme_boost_union/recommendationdetailsmodal
+ * @copyright  2026 Alexander Bias <bias@alexanderbias.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -28,7 +27,7 @@ import Templates from 'core/templates';
 import {getString} from 'core/str';
 
 const SELECTORS = {
-    DETAILS: '[data-action="courselisting-details"]',
+    DETAILS: '[data-action="recommendation-details"]',
 };
 
 /**
@@ -41,7 +40,7 @@ export const init = () => {
 };
 
 /**
- * Register snippet related event listeners.
+ * Register recommendation related event listeners.
  *
  * @method registerListenerEvents
  */
@@ -62,19 +61,16 @@ const registerListenerEvents = () => {
  * @param {object} element
  */
 const buildModal = async(element) => {
-
     // Prepare data for modal.
     const data = {
         title: element.getAttribute('data-title'),
         summary: element.getAttribute('data-summary'),
-        coursecontacts: element.getAttribute('data-coursecontacts'),
-        customfields: element.getAttribute('data-customfields'),
+        description: element.getAttribute('data-description'),
     };
 
     await ModalCancel.create({
         title: data.title,
-        body: Templates.render('theme_boost_union/courselistingdetailsmodal', data),
-        large: true,
+        body: Templates.render('theme_boost_union/recommendationsdetailsmodal', data),
         buttons: {
             'cancel': getString('closebuttontitle', 'moodle'),
         },